Sponsoring

+++ wp-content/themes/Cours-formation-wordpress/functions.php
@@ -67,7 +67,7 @@ function cours_wordpress_render_sponso_box(){
 }
 
 function cours_wordpress_save_sponso($post_id){
-	if(array_key_exists('cours_sponso', $_POST)){
+	if(array_key_exists('cours_sponso', $_POST)&& current_user_can( 'edit_post',$post_id )){
 		if($_POST['cours_sponso']==='0'){
 			delete_post_meta($post_id,'cours_sponso');
 		}
